Incorporated theme (Home/Slider page and Navbar component. Added starter readme file.  Added buttons [product search and home].
