+++ README.md
@@ -1,3 +1,91 @@
-# Good-Food-For-You
+# Project Three:  Good Foods & Meal Planning
 
-## hello
+  [![license](https://img.shields.io/badge/license-MIT-blue.svg)](https://choosealicense.com/licenses/mit/)
+
+***
+
+  ## Description:
+  Meal plan and groceries... Update Later. 
+  
+***
+  ## Table of Contents:
+  1.  [Description](#description)
+  2.  [Demonstration](#demonstration)
+  3.  [Installation](#installation)
+  4.  [Screenshots](#screenshots)
+  5.  [Technologies](#technologies)
+  6.  [Usage](#usage)
+  7.  [License](#license)
+  8.  [Contributing](#contributing)
+  9.  [Tests](#tests)
+  10.  [Questions](#questions)
+
+***
+  ## Demonstration:
+  [Deployed website link]()
+
+***
+  ## Installation:
+  To use this application, the user can fork or clone.  Then run `npm install`. See the technologies section for the application dependencies.  In addition, the user will need to connect to an active database server for full functionality.  However, any users may freely use this deployed application at the above deploy application website.
+
+***
+  ## Screenshots:
+  ### (Deployed Application)
+  ![Deployed Application]()
+
+  ### (User Login)
+  ![User Login]()
+
+  ### (Meal Plan Search)
+  ![Meal Plan Search]()
+
+  ### (Grocery List)
+  ![Grocery List]()
+
+***
+  ## Technologies
+    - Node.js
+    - Express.js
+    - MongoDB
+    - React
+    - Dotenv
+    - Bcryptjs
+    - Materialize
+
+  ## Usage:
+  This application utilizes several web and database technologies... Update Later...  
+   
+***
+  ## License:
+  This project falls under the MIT License.  The full documentation for this license can be found at [MIT Full Documentation](https://choosealicense.com/licenses/mit).
+
+  Below is an excerpt of the MIT License.
+  <br>
+  Copyright <YEAR> <COPYRIGHT HOLDER>
+    
+    Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+    The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+  ## Contributing:
+  To contribute to this project, please send an email.  
+  <br>
+  The contributors for this project are:
+    - Bruce Mowrey
+    - Norman Cerros
+    - Leighton Jenkins
+    - D. Javan Worthy
+
+***
+  ## Tests:
+  There are some test for this application.  The data/communication routes must be verified before this application can be properly deployed.  In addition, there must be an operational database  and it must be connected to the application to receive the full functionality.  
+
+***
+  ## Questions:
+  - The GitHub profiles for this project are:
+    - https://github.com/bmowrey729
+    - https://github.com/ncerros
+    - https://github.com/LD-Jenkins
+    - https://github.com/djavanw
+ 
+  - If there are any questions, please contact Leighton at leightonjenkins84@gmail.com.
