+++ client/src/App.css
@@ -1,0 +1,3 @@
+body {
+    background:  rgb(194, 243, 209);
+}
