+++ client/src/App.js
@@ -2,6 +2,7 @@ import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import Product from "./pages/Product";
 import Details from "./pages/Details";
+import Home from "./pages/Home";
 //import NoMatch from "./pages/NoMatch";
 import Nav from "./components/Nav";
 
@@ -12,7 +13,10 @@ function App() {
       <div>
         <Nav />
         <Switch>
-          <Route exact path={["/", "/product"]}>
+          <Route exact path={["/", "/slider"]}>
+            <Home />
+          </Route>
+          <Route exact path="/product">
             <Product />
           </Route>
         <Route exact path="/product/:id">
