+++ client/src/components/DeleteBtn/index.js
@@ -3,10 +3,11 @@ import "./style.css";
 
 // The ...props means, spread all of the passed props onto this element
 // That way we don't have to define them all individually
+
 function DeleteBtn(props) {
   return (
     <span className="delete-btn" {...props} role="button" tabIndex="0">
-      ✗ 
+      <i className="small material-icons">clear</i>
     </span>
   );
 }
