+++ client/src/components/Nav/index.js
@@ -1,24 +1,24 @@
-import React from "react";
+import React, { Component } from 'react'
 
-function Nav() {
-  return (
-    <nav>
-    <div class="nav-wrapper">
-      <a href="/" class="brand-logo">Good Foods</a>
-      <ul id="nav-mobile" class="right hide-on-med-and-down">
-        <li><a href="/">Grocery Product List</a></li>
-        
-      </ul>
-    </div>
-  </nav>
-  );
+/*eslint-disable */
+class Nav extends Component {
+    render() {
+        return (
+          <div>
+            <nav>
+              <div class="nav-wrapper #81c784 green lighten-2">
+                <a href="#" class="brand-logo left">Good Foods</a>
+                <ul id="nav-mobile" class="right">
+                  <li><a href="/" className="waves-effect waves-light btn #66bb6a green lighten-1">HOME</a></li>
+                  <li><a href="/product" className="waves-effect waves-light btn #66bb6a green lighten-1">PRODUCT SEARCH</a></li>
+                  <li><a href="#" className="waves-effect waves-light btn #66bb6a green lighten-1">LOGIN</a></li>
+                  <li><a href="#" className="waves-effect waves-light btn #66bb6a green lighten-1">SIGNUP</a></li>
+                </ul>
+              </div>
+            </nav>
+          </div>
+        )
+    }
 }
 
 export default Nav;
-
-
-    // <nav className="navbar navbar-expand-lg navbar-dark bg-primary">
-    //   <a className="navbar-brand" href="/">
-    //     Good Food & Meal Planning
-    //   </a>
-    // </nav>
